Allow providing a list of known inhibitors
Add a variable to allow providing a list of keys of known inhibitors,
these known inhibitors will be ignored when setting leapp_inhibitors
and considering the value for upgrade_inhibited.

Example usage, here VDO related inhibitors are listed as known
and they will not cause the playbook to fail on any host:

```
    leapp_known_inhibitors:
      # vdo check
      - 2f9802dc91315806c7cdc3c18d7b74f2a2383285
      # vdo package
      - 429a99e13b19a7eebadbb8cb35233d8119bcf255
  tasks:
    - name: Generate preupgrade analysis report
      ansible.builtin.import_role:
        name: infra.leapp.analysis

    - name: Fail if unknown inhibitors found
      fail:
        msg: Unknown inhibitors found, see the logs for details.
      when: upgrade_inhibited
```

This is helpful when dealing with a large number of hosts and where
remediations for the known inhibitors are available, for any new hosts
only unknown or unhandled inhibitors would cause the playbook to fail
and thus the need to investigate in detail.

Resolves: #123