fix post upgrade task failing on package conflicts
The post-upgrade task which updates the system packages failed when
running upgrade with custom repositories. Since leapp does not update the
repositories in /etc/yum.repos.d when upgrading with custom
repositories, the post-upgrade update task is now skipped in the case
when the leapp_upgrade_type is custom and when the
local_repos_post_upgrade are not defined.

Jira: RHELMISC-21816