test: add test for cifs remediation
Add a test to trigger and verify the cifs remediation, and test
playbooks for 7to8, 8to9, and 9to10 in order to run remediations
for those cases.

This adds a sort of test framework for doing test setup and test
verify, by adding `tests/tasks/setup/` and `tests/tasks/verify/`
sub-directories with task files for those purposes.  We currently
use it only for doing remediation setup and verify, but I could
imagine using it for other purposes, or adding a `cleanup/`
sub-directory for cleanup tasks.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>